# API-Gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing LinguaLeap System resources",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/languages/exercise/{exercise_id}/submit": {
            "post": {
                "description": "Submit an exercise for evaluation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Submit an exercise",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Exercise ID",
                        "name": "exercise_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Submit Exercise",
                        "name": "submit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/learn.GetQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.GetQuestionResponse"
                        }
                    }
                }
            }
        },
        "/languages/exercises/{language_code}": {
            "get": {
                "description": "Get exercises for a language",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Get exercises",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.GetQuestionResponse"
                        }
                    }
                }
            }
        },
        "/languages/lesson/{lesson_id}": {
            "get": {
                "description": "Get details for a specific lesson",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Get lesson details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson ID",
                        "name": "lesson_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.GetLessonResponse"
                        }
                    }
                }
            }
        },
        "/languages/lesson/{lesson_id}/complete": {
            "post": {
                "description": "Mark a lesson as completed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Complete a lesson",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson ID",
                        "name": "lesson_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Complete Lesson",
                        "name": "complete",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/learn.GetLessonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.GetLessonResponse"
                        }
                    }
                }
            }
        },
        "/languages/lessons/{language_code}": {
            "get": {
                "description": "Get lessons for a language",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Get lessons",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.GetLessonResponse"
                        }
                    }
                }
            }
        },
        "/languages/start": {
            "post": {
                "description": "Start learning a language",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Start learning a language",
                "parameters": [
                    {
                        "description": "Start Learning Language",
                        "name": "start",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/learn.CreateLanguageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.CreateLanguageResponse"
                        }
                    }
                }
            }
        },
        "/languages/vocabulary/{language_code}": {
            "get": {
                "description": "Get vocabulary for a language",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Get vocabulary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.GetVocabularyResponse"
                        }
                    }
                }
            }
        },
        "/languages/vocabulary/{language_code}/add": {
            "post": {
                "description": "Add new vocabulary to a language",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "languages"
                ],
                "summary": "Add vocabulary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Add Vocabulary",
                        "name": "vocabulary",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/learn.AddVocabularyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/learn.AddVocabularyResponse"
                        }
                    }
                }
            }
        },
        "/progress/achievements": {
            "get": {
                "description": "Get the user's achievements",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get achievements",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.AchievementsResponse"
                        }
                    }
                }
            }
        },
        "/progress/daily": {
            "get": {
                "description": "Get the user's daily progress",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get daily progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.DailyProgressResponse"
                        }
                    }
                }
            }
        },
        "/progress/delete-goal/{goal_id}": {
            "delete": {
                "description": "Delete an existing goal",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Delete a goal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Goal ID",
                        "name": "goal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.GoalResponse"
                        }
                    }
                }
            }
        },
        "/progress/goals": {
            "get": {
                "description": "Get the user's goals",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get goals",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.GoalsResponse"
                        }
                    }
                }
            }
        },
        "/progress/leaderboard/{language_code}": {
            "get": {
                "description": "Get the leaderboard for a specific language",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get leaderboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.LeaderboardResponse"
                        }
                    }
                }
            }
        },
        "/progress/monthly": {
            "get": {
                "description": "Get the user's monthly progress",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get monthly progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.MonthlyProgressResponse"
                        }
                    }
                }
            }
        },
        "/progress/set-goal": {
            "post": {
                "description": "Set a new goal for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Set a goal",
                "parameters": [
                    {
                        "description": "Set Goal",
                        "name": "goal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/progres.Goal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.GoalResponse"
                        }
                    }
                }
            }
        },
        "/progress/skills/{language_code}": {
            "get": {
                "description": "Get the user's skills for a specific language",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get skills",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.SkillsResponse"
                        }
                    }
                }
            }
        },
        "/progress/statistics/{language_code}": {
            "get": {
                "description": "Get the user's statistics for a specific language",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.StatisticsResponse"
                        }
                    }
                }
            }
        },
        "/progress/update-goal/{goal_id}": {
            "put": {
                "description": "Update an existing goal for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Update a goal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Goal ID",
                        "name": "goal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Goal",
                        "name": "goal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/progres.UpdateGoalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.GoalResponse"
                        }
                    }
                }
            }
        },
        "/progress/weekly": {
            "get": {
                "description": "Get the user's weekly progress",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get weekly progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.WeeklyProgressResponse"
                        }
                    }
                }
            }
        },
        "/progress/{language_code}": {
            "get": {
                "description": "Get the user's progress for a specific language",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get language progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.ProgressResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "learn.AddVocabularyRequest": {
            "type": "object",
            "properties": {
                "language_code": {
                    "description": "Til kodi, masalan, 'en', 'uz'",
                    "type": "string"
                },
                "translation": {
                    "description": "So'z tarjimasi",
                    "type": "string"
                },
                "word": {
                    "description": "So'z",
                    "type": "string"
                }
            }
        },
        "learn.AddVocabularyResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Yaratilgan leksika yozuving UUID si",
                    "type": "string"
                }
            }
        },
        "learn.CreateLanguageRequest": {
            "type": "object",
            "properties": {
                "emoji": {
                    "description": "Til flag emojisi",
                    "type": "string"
                },
                "language_code": {
                    "description": "Til kodi, masalan, 'en', 'uz'",
                    "type": "string"
                },
                "name": {
                    "description": "Til nomi, masalan, 'Ingliz tili', 'O'zbek tili'",
                    "type": "string"
                }
            }
        },
        "learn.CreateLanguageResponse": {
            "type": "object",
            "properties": {
                "language": {
                    "description": "Yaratilgan til haqida ma'lumot",
                    "allOf": [
                        {
                            "$ref": "#/definitions/learn.Language"
                        }
                    ]
                }
            }
        },
...
        "learn.GetLessonRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Darsning UUID si",
                    "type": "string"
                }
            }
        },
        "learn.GetLessonResponse": {
            "type": "object",
            "properties": {
                "lesson": {
                    "description": "Dars haqida ma'lumot",
                    "allOf": [
                        {
                            "$ref": "#/definitions/learn.Lesson"
                        }
                    ]
                }
            }
        },
        "learn.GetQuestionRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Savolning UUID si",
                    "type": "string"
                }
            }
        },
        "learn.GetQuestionResponse": {
            "type": "object",
            "properties": {
                "question": {
                    "description": "Savol haqida ma'lumot",
                    "allOf": [
                        {
                            "$ref": "#/definitions/learn.Question"
                        }
                    ]
                }
            }
        },
        "learn.GetVocabularyResponse": {
            "type": "object",
            "properties": {
                "vocabulary": {
                    "description": "Leksika yozuvi haqida ma'lumot",
                    "allOf": [
                        {
                            "$ref": "#/definitions/learn.Vocabulary"
                        }
                    ]
                }
            }
        },
...
        "progres.AchievementsResponse": {
            "type": "object",
            "properties": {
                "achievements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.Achievement"
                    }
                }
            }
        },
...
        "progres.DailyProgressResponse": {
            "type": "object",
            "properties": {
                "progress": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.DailyProgress"
                    }
                }
            }
        },
        "progres.Goal": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Maqsad yaratilgan sana",
                    "type": "string"
                },
                "current_value": {
                    "description": "Hozirgi qiymat",
                    "type": "integer"
                },
                "end_date": {
                    "description": "Maqsad tugash sanasi",
                    "type": "string"
                },
                "goal_type": {
                    "description": "Maqsad turi (masalan, 'daily_xp', 'weekly_lessons', va hokazo)",
                    "type": "string"
                },
                "id": {
                    "description": "Maqsad ID'si",
                    "type": "string"
                },
                "language_code": {
                    "description": "Til kodi (masalan, 'fr', 'es', va hokazo)",
                    "type": "string"
                },
                "start_date": {
                    "description": "Maqsad boshlanish sanasi",
                    "type": "string"
                },
                "target_value": {
                    "description": "Maqsad uchun kerakli qiymat",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "Maqsad yangilangan sanaa",
                    "type": "string"
                },
                "user_id": {
                    "description": "Foydalanuvchi ID'si",
                    "type": "string"
                }
            }
        },
        "progres.GoalResponse": {
            "type": "object",
            "properties": {
                "goal": {
                    "$ref": "#/definitions/progres.Goal"
                }
            }
        },
        "progres.GoalsResponse": {
            "type": "object",
            "properties": {
                "goals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.Goal"
                    }
                }
            }
        },
...
        "progres.LeaderboardResponse": {
            "type": "object",
            "properties": {
                "leaderboards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.Leaderboard"
                    }
                }
            }
        },
...
        "progres.MonthlyProgressResponse": {
            "type": "object",
            "properties": {
                "progress": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.MonthlyProgress"
                    }
                }
            }
        },
        "progres.ProgressResponse": {
            "type": "object",
            "properties": {
                "daily_progress": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.DailyProgress"
                    }
                },
                "monthly_progress": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.MonthlyProgress"
                    }
                },
                "weekly_progress": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.WeeklyProgress"
                    }
                }
            }
        },
...
        "progres.SkillsResponse": {
            "type": "object",
            "properties": {
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.Skill"
                    }
                }
            }
        },
...
        "progres.StatisticsResponse": {
            "type": "object",
            "properties": {
                "statistics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.Statistic"
                    }
                }
            }
        },
        "progres.UpdateGoalRequest": {
            "type": "object",
            "properties": {
                "goal": {
                    "$ref": "#/definitions/progres.Goal"
                },
                "goal_id": {
                    "type": "string"
                }
            }
        },
...
        "progres.WeeklyProgressResponse": {
            "type": "object",
            "properties": {
                "progress": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.WeeklyProgress"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}